drivers: move GPS drivers and include header
Move gps.h into /include/drivers/gps
Move gps_sim and nrf9160_gps into own /drivers/gps folder.
