+++ drivers/gps/nrf9160_gps/CMakeLists.txt
