+++ samples/nrf9160/lwm2m_client/src/lwm2m/lwm2m_location.c
@@ -5,7 +5,7 @@
  */
 
 #include <zephyr.h>
-#include <gps.h>
+#include <drivers/gps.h>
 #include <stdio.h>
 #include <net/lwm2m.h>
